--- a/go.sh
+++ b/go.sh
@@ -3,16 +3,15 @@ set -ea
 
 SCALE_OPENRESTY=${SCALE_OPENRESTY:-1}
 SCALE_APP=${SCALE_APP:-1}
-PROJECT=${PROJECT:-$(basename "${PWD}" | sed 's/[\w.-]//g')}
 
 touch acme.json
 chmod 600 acme.json
 
 echo "Building services ..."
-docker-compose -p "${PROJECT}" -f "${DOCKER_COMPOSE_FILE:-docker-compose.yml}" build
+docker-compose build
 
 echo "Starting services ..."
-docker-compose -p "${PROJECT}" -f "${DOCKER_COMPOSE_FILE:-docker-compose.yml}" up -d \
+docker-compose up -d \
   --remove-orphans \
   --scale openresty="$SCALE_OPENRESTY" \
   --scale php-fpm="$SCALE_APP"