… clean.sh
… docker-compose.full.yml
… docker-compose.stateless.yml
… docker-compose.yml
… docs/installation.md
… scripts/clean-content.sh
… scripts/clean-persistence.sh
… scripts/install-deps.sh
… scripts/update-repos.sh
… scripts/watch.sh
… update.sh
… wait.sh