Ref: https://jira.greenpeace.org/browse/PLANET-5483
Tries to do things for: https://jira.greenpeace.org/browse/PLANET-4752

> Investigate ways of pre-installing node/npm to our dev images


## Chchchch-chch-changes

### Node container
- A basic [node container](https://hub.docker.com/_/node/) is used to run all npm operations (build/install/watch/etc.).  
- `node_modules` folders are excluded from mounting, which seems to help (MacOS) performances a bit.

### No Elastic for default build
Working toward a lighter default build, elastic container is not built by default.

### Variables changes
`PROJECT` and `DOCKER_COMPOSE_FILE`are replaced with their native counterpart `COMPOSE_PROJECT_NAME` and `COMPOSE_FILE`
Those are written in a `.env` file natively used by docker-compose; any command executed from this folder will use those values, which makes working with docker-compose a lot easier.  
Old variables can be removed from the docker-compose commands, and you don't have to run `eval $(make env)` in your shell sessions to work with it.  `COMPOSE_ENV` is dropped.
To change env variables, edit `.env` file or run `COMPOSE_FILE="myfile.yml" make env`.  

### Local scripts
Added a bunch of scripts, to replace some included in the docker repository, and some to split a bit more make operations.
```
scripts                 
├── build-assets.sh      // Builds js/css assets
├── clean-content.sh     // Removes default content folder
├── clean-persistence.sh // Removes persistence folder
├── install-deps.sh      // Install composer and npm dependencies
├── update-repos.sh      // Update main theme and gutenberg blocks depos
└── watch.sh             // runs npm watch (options --theme-only --plugin-only)
```

### Update command
For devs not working on the main repos it can be easy to forget to get up to date with our changes.  
`make update` updates main repos and rebuilds assets.

## Todo next
- rsync operations take up to a minute and might not be needed after all (problem with ci?)
- replace npm with pnpm (or other?) for install/build, to try to speed up this step
- I think we can run dev env without traefik and openresty, by using php builtin webserver and a [handwritten router](https://stackoverflow.com/a/38926070). I didn't succeed at this, mostly because I don't seem to understand port mapping and declared networks.
- remove some log errors by maybe adding empty file `/app/source/public/wp-content/wflogs/rules.php`

# Test

- Save your work
  - **Commit the things you're working on**, it will probably disappear from your disk
- Clean your system
  - clean planet4 things `make clean` (removing `defaultcontent` is optional)
  - (optional) clean all docker things `docker system prune -a` (it will remove _everything_ docker has, including images previously downloaded and anything that doesn't run)
- Check if you have a `.env` file - if you don't, everything's fine - if you do, please add `COMPOSE_FILE="docker-compose.yml"` and `COMPOSE_PROJECT_NAME="planet4dockercompose"`
- Run `make dev`
- Be patient
- Report crashes

_`make dev` time, (without images/content/dependencies downloads, using cache)  
(downloads will significantly extend build time depending on your internet connection speed)_
| laptop | system | time (without downloads) |
| --- | --- | --- |
| x250 i5 8G | ubuntu 20.04 | 6m40 |
| macbookair 2020 i5 8G | macos catalina 10.15.7 | 8m50 |