@comzeradd 2021-02-08
Once this is merged, we should remove the duplicate scripts from [base-fork](https://github.com/greenpeace/planet4-base-fork/tree/develop/tasks/other), right?