@lithrel 2021-02-09
> Once this is merged, we should remove the duplicate scripts from [base-fork](https://github.com/greenpeace/planet4-base-fork/tree/develop/tasks/other), right?

Yep that's the plan, I just need to be sure it's not used elsewhere